requirements.txt
django
djangorestframework
djoser
django-split-settings
pyyaml
uritemplate
httpx